(finally) moved database load and save  methods to the DatabaseUtils class, removed legacy database load/save methods
